event_generator/requirements.txt
kafka-python==2.0.2
pandas==2.0.3
numpy==1.24.3
Faker==19.3.0
python-dateutil==2.8.2
requests==2.31.0 